+++ changelog.md
@@ -29,6 +29,7 @@
 
 ### Added
 
+- New nested NER trainable `nested_ner` pipeline component
 - Now possible to provide regex flags when using the RegexMatcher
 - New `ContextualMatcher` pipe, aiming at replacing the `AdvancedRegex` pipe.
 - New `as_ents` parameter for `eds.dates`, to save detected dates as entities